requirements.txt
streamlit
pandas
plotly-express
numpy
scikit-learn